Breakdown Post component into smaller components
